+++ Domain/RDD.Domain/IRepository.cs
@@ -8,5 +8,6 @@ namespace Rdd.Domain
         void Add(TEntity entity);
         void AddRange(IEnumerable<TEntity> entities);
         void Remove(TEntity entity);
+        void DiscardChanges(TEntity entity);
     }
 }